requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
yagmail>=0.15.0

# Email sending (choose one):
# sendgrid>=6.11.0  # Recommended for production (uncomment to use)

